+++ controllers/testRouterController.js
@@ -23,6 +23,16 @@ const prisma = new PrismaClient();
 
 //
 
+const { createClient } = require("@supabase/supabase-js");
+
+// Create a single supabase client for interacting with your database
+const supabase = createClient(
+  "http://localhost:54321",
+  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZS1kZW1vIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImV4cCI6MTk4MzgxMjk5Nn0.EGIM96RAZx35lJzdJsyH-qQwv8Hdp7fsn3W0YpN81IU"
+);
+const fs = require("fs");
+const path = require("path");
+
 // const multer = require("multer");
 // const storage = multer.diskStorage({
 //   destination: (req, file, cb) => {
@@ -42,47 +52,79 @@ const prisma = new PrismaClient();
 // const upload = multer({ storage }).single("profileImage");
 
 const handleTestGet = tryCatchMiddleware(async (req, res, next) => {
-  // let availableClients = await handlePrismaGetSingleData("client");
+  const { dataaaa, errorrrr } = await supabase.storage.createBucket("uploads", {
+    public: true,
+    // allowedMimeTypes: ["image/png", ],
+    // fileSizeLimit: 5000,
+  });
 
-  responseSend(res, { fail: null });
+  // console.log(data, error);
+
+  // const { data, error } = await supabase.storage.listBuckets();
+
+  // const filePath = path.join(
+  //   __dirname,
+  //   "../public/uploads/profilePhotos/54865.jpeg"
+  // );
+  // const fileData = fs.readFileSync(filePath);
+
+  // console.log("filePath ----", path.extname(filePath));
+
+  // const avatarFile = fileData;
+  // const { data, error } = await supabase.storage
+  //   .from("avatars")
+  //   .upload(`${"filename" + path.extname(filePath)}`, avatarFile, {
+  //     cacheControl: "3600",
+  //     upsert: true,
+  //   });
+
+  // const { datas } = supabase.storage
+  //   .from("avatars")
+  //   .getPublicUrl("filename.jpeg");
+  // console.log("url of image ------- ", datas);
+
+  // console.log("All buckets ------- ", data);
+  // console.log("All buckets ------- ", error);
+
+  // responseSend(res, { data: data, error: error });
+
+    const filePath = path.join(
+      __dirname,
+      "../public/uploads/profilePhotos/54865.jpeg"
+    );
+  const fileData = fs.readFileSync(filePath);
+
+  const fileName = `filename${path.extname(filePath)}`;
+
+  const { data, error } = await supabase.storage
+    .from("uploads")
+    .upload(fileName, fileData, {
+      cacheControl: "3600",
+      upsert: true,
+    });
+
+  if (error) {
+    // Handle error
+    responseSend(res, { error });
+    return;
+  }
+
+  const publicURL = await supabase.storage
+    .from("avatars")
+    .getPublicUrl(fileName);
+
+  console.log("URL of image:", publicURL);
+
+  responseSend(res, { data, publicURL });
 });
 
 const handleTestPost = tryCatchMiddleware(async (req, res, next) => {
-  const { userId, ...profileData } = req.body;
-
-  let data = await prisma.profile.upsert({
-    create: {
-      ...profileData,
-      userId,
-    },
-    update: {
-      ...profileData,
-    },
-    where: {
-      userId,
-    },
-    include: {
-      user: true,
-    },
-  });
-
-  function excludePasswordAndID(user, keys) {
-    for (let key of keys) {
-      delete user[key];
-      delete user.user[key];
-    }
-    return user;
-  }
-  const userWithoutPassword = excludePasswordAndID(data, ["password", "id"]);
-
-  responseSend(res, userWithoutPassword);
+  responseSend(res, { fail: null });
 });
 
 const handleTestImageUpload = tryCatchMiddleware(async (req, res, next) => {
-
-  let data = {"status":"uploaded"};
+  let data = { status: "uploaded" };
   responseSend(res, data);
-
 });
 
 module.exports = {