+++ prisma/migrations/20230507083303_init/migration.sql
@@ -1,2 +1,0 @@
--- AlterTable
-ALTER TABLE "Client" ALTER COLUMN "redirectUris" DROP NOT NULL;