+++ prisma/migrations/20230507091723_init/migration.sql
@@ -1,9 +1,0 @@
-/*
-  Warnings:
-
-  - The `redirectUris` column on the `Client` table would be dropped and recreated. This will lead to data loss if there is data in the column.
-
-*/
--- AlterTable
-ALTER TABLE "Client" DROP COLUMN "redirectUris",
-ADD COLUMN     "redirectUris" TEXT[] DEFAULT ARRAY[]::TEXT[];