+++ prisma/migrations/20230509064642_init/migration.sql
@@ -1,2 +1,0 @@
--- AlterTable
-ALTER TABLE "Profile" ADD COLUMN     "profilePhoto" TEXT DEFAULT 'https://storage.googleapis.com/bucket-playground-tal/user-profiles/user.webp';