+++ prisma/migrations/20230510085934_init/migration.sql
@@ -1,3 +1,0 @@
--- AlterTable
-ALTER TABLE "Profile" ADD COLUMN     "deleted" BOOLEAN NOT NULL DEFAULT false,
-ADD COLUMN     "locked" BOOLEAN NOT NULL DEFAULT false;