+++ CHANGELOG.md
@@ -3,21 +3,25 @@
 ## [1.0.0]
 
 ### First upload
-- Encryption, response, error handling middleware 
+
+- Encryption, response, error handling middleware
 
 ## [1.1.0]
 
 ### Device routes added
-- device routes added, changes and modifications 
+
+- device routes added, changes and modifications
 
 ## [1.2.0]
 
 ### Device routes added
+
 - auth routes added, changes and modifications, oAuth implemented
 
 ## [1.3.0]
 
 ### encryption and oAuth updated
+
 - oauth works for both encrypted and non-encrypted requests
 - oAuth routes updated for json body acceptance
 - minor fix up overall the code
@@ -26,13 +30,19 @@
 
 - authorization token checking method changed
 - changes in routes
+
 ## [1.5.0]
 
 - SQL related
+
 ## [1.5.0]
 
 - SQL related and oauth route changed also admin route is updated
+
 ## [1.7.0]
 
 - admin route and functions updated
 
+## [1.7.0]
+
+- added supabase support